+++ app/src/test/java/com/android/sample/HomeViewModelTest.kt
@@ -1,5 +1,9 @@
 package com.android.sample
 
+import com.android.sample.repositories.FakeHomeRepository
+import com.android.sample.repositories.HomeRepository
+import com.android.sample.repositories.HomeUiState
+import com.android.sample.repositories.HomeViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.*
